pyproject.toml
[tool.poetry]
name = "rico"
version = "0.1.0"
description = "Transient alert generation for Argus and Evryscope"
authors = ["Hank Corbett"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.18.0"
fastavro = "^1.4.12"
pyarrow = "^8.0.0"
confluent-kafka = "^2.1.1"
fastapi = "^0.95.2"
rich = "^13.3.5"
click = "^8.1.3"
black = "^23.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
mypy = "^0.931"
isort = "^5.10.1"
flake8 = "^4.0.1"
Sphinx = "^7.0.1"
pre-commit = "^3.3.2"


[tool.poetry.scripts]
eatr = "eatr.cli:main"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = "true"
force_grid_wrap = 0
use_parentheses = "true"
line_length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

